+++ src/app/app.tsx
@@ -6,7 +6,6 @@ import {LoginPage} from '../pages/login-page/login-page.tsx';
 import {MainPage} from '../pages/main-page/main-page.tsx';
 import {NotFound404Page} from '../pages/not-found-404-page/not-found-404-page.tsx';
 import {OfferPage} from '../pages/offer-page/offer-page.tsx';
-import {PropertyPage} from '../pages/property-page/property-page.tsx';
 import {RoutePath} from '../shared/enums/routes.ts';
 
 export const App: FC = () => {
@@ -21,7 +20,6 @@ export const App: FC = () => {
         <Route path={RoutePath.OfferPage} element={<OfferPage/>}>
           <Route path=":id" element={<OfferPage/>}/>
         </Route>
-        <Route path={RoutePath.PropertyPage} element={<PropertyPage/>}/>
         <Route path={RoutePath.Page404} element={<NotFound404Page/>}/>
         <Route path="*" element={<NotFound404Page/>}/>
       </Routes>
