+++ src/components/map-widget/model/types.ts
@@ -1,5 +1,5 @@
 import {Key, ReactNode} from 'react';
-import {Coordinates} from '../../../shared/types/coordinates.ts';
+import {Coordinates} from '../../../shared/entities/coordinates/coordinates.ts';
 
 export type PointOnMap = {
   id: Key;
