+++ src/pages/favorites-page/favorite-offer-card-list/favorite-offer-card-list.tsx
