+++ src/pages/favorites-page/favorite-offer-card-list/favorite-offer-card.tsx
@@ -1,10 +1,10 @@
 import {FC} from 'react';
 import {Link} from 'react-router-dom';
 import {Offer} from '../../../shared/entities/offer/types.ts';
-import {changeFavoriteStatus} from '../../../slices/favorites-page-slice.ts';
 import {RoutePath} from '../../../shared/enums/routes.ts';
 import {useAppDispatch} from '../../../shared/redux-helpers/typed-hooks.ts';
 import {FavoriteStatus} from '../../../shared/server-interaction/constants.ts';
+import {changeFavoriteStatus} from '../../../slices/favorites-page-slice.ts';
 
 type FavoriteOfferCardProps = {
   offer: Offer;
