+++ src/pages/offer-page/offer-page.tsx
@@ -1,15 +1,15 @@
 import {FC, useCallback, useEffect} from 'react';
 import {Navigate, useNavigate, useParams} from 'react-router-dom';
-import {Offer, OfferExtendedDto} from '../../shared/entities/offer/types.ts';
+import {MapWidget} from '../../components/map-widget/map-widget.tsx';
+import {PointOnMap} from '../../components/map-widget/model/types.ts';
 import {OfferCardList} from '../../components/offer-card-list/offer-card-list.tsx';
-import {changeFavoriteStatus} from '../../slices/favorites-page-slice.ts';
-import {loadNearbyOffers, loadOffer} from '../../slices/offer-page-slice.ts';
+import {Spinner} from '../../components/spinner/spinner.tsx';
+import {Offer, OfferExtendedDto} from '../../shared/entities/offer/types.ts';
 import {RoutePath} from '../../shared/enums/routes.ts';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
 import {FavoriteStatus} from '../../shared/server-interaction/constants.ts';
-import {PointOnMap} from '../../components/map-widget/model/types.ts';
-import {MapWidget} from '../../components/map-widget/map-widget.tsx';
-import {Spinner} from '../../components/spinner/spinner.tsx';
+import {changeFavoriteStatus} from '../../slices/favorites-page-slice.ts';
+import {loadNearbyOffers, loadOffer} from '../../slices/offer-page-slice.ts';
 import {CommentsList} from './comments-list.tsx';
 import {FeedbackForm} from './feedback-form.tsx';
 import {OfferPageWrapper} from './offer-page-wrapper.tsx';
