+++ src/shared/entities/coordinates/coordinates.ts
